requirements.txt
botbuilder-core>=4.14.0
botbuilder-integration-aiohttp>=4.14.0
aiohttp>=3.9.0
openai>=1.0.0
anthropic>=0.34.0
msal>=1.28.0
httpx>=0.27.0
python-dotenv>=1.0.0
